updated commit containing profile picture upload to s3 bucket and url store in database updated with qr code generation for orders and url store in the database contains functioning implementation of mapping api and complete services for preorder

+++ checkit/src/main/java/com/example/checkit/dtos/AddressDto.java
@@ -3,13 +3,15 @@ package com.example.checkit.dtos;
 import lombok.Data;
 import lombok.experimental.Accessors;
 
+import java.math.BigDecimal;
+
 @Data
 @Accessors(chain = true)
 public class AddressDto {
 
     protected String addressName;
 
-    protected  Long addressLongitude;
+    protected BigDecimal addressLongitude;
 
-    protected Long addressLatitude;
+    protected BigDecimal addressLatitude;
 }
